(1209) feat: 게시글 동적 정렬 및 카테고리 필터링 구현 - QueryDSL 동적 정렬 기능 추가 (최신순/조회수순/제목순) - 카테고리 필터링 기능 구현 - 검색어 + 카테고리 + 정렬 복합 조건 지원 - 게시글 조회 시 조회수 자동 증가 - 성능 최적화를 위한 인덱스 추가 (created_at, view_count, category_id)

+++ src/main/java/com/cho/board/post/dtos/PostDetailResponse.java
@@ -13,7 +13,7 @@ public class PostDetailResponse {
     private Long id;
     private String title;
     private String content;
-    private int viewCount;
+    private Long viewCount;
 
     private Long authorId;
     private String authorName;
